feat: add plumbing to use theme for colors and status bar (#24)
### Description

As the title.

### Test plan

Tested with the example app that custom colors and status bar colors are
applied. (Below with backgroundPrimary = yellow, isDark = true):
![Simulator Screenshot - iPhone 15 Pro - 2025-01-28 at 12 26
05](https://github.com/user-attachments/assets/3c3bf814-e2a1-4ef4-bcf5-95e8b4d3df35)



### Related issues

- Related to RET-1280

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)