feat: expose useWallet hook from the public folder (#25)
### Description
This is a proposal for the initial structure of the exposed data hooks
that provides better UX for builders by abstracting selectors. Its
intention is to hide the complexity of our redux state and provide as
much useful info for builders as possible. This will also give us an
ability to optimize existing selectors (e.g. by memoizing deps) without
re-writing them and without pushing that memoization management on the
builders.
Currently, this only exports `useWallet` to gather some feedback on
whether this approach is viable and if we should proceed with it.

There were some hurdles with the module resolution as `useWallet` is the
first file from `src/public` folder that uses absolute internal paths
like `src/*` for its imports. Including this in the build of
`@interaxyz/mobile` module involves including a lot of internal files
that have to be properly resolved.

Reasoning for most of the code is explained in the comments below.
Please, feel free to share your feedback!

Currently suggested proposal of exposed data hooks looks like this:

```tsx
function Component(props: Props) {
  const { address, tokens } = useWallet()
  
  return (
    <View>
      <Text>Wallet address: {address}</Text>
      
      {tokens.map(tokenInfo => (
        <Text>{tokenInfo.symbol}</Text>
      ))}
    </View>
  )
}
```

### Test plan
N/A

### Related issues
Relates to RET-1280

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)