fix: remove flipper from store setup (#28)
### Description

This change removes the need to include `react-native-flipper` and
`redux-flipper` in end applications. Expo does not use flipper anyway.

### Test plan

n/a

### Related issues

- Related to RET-1289

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)