chore: update colors of transaction feed and points history icons (#6450)
### Description

This is planned for the transaction feed work but it also affects beefy,
so I thought I'd just do it now. Kayla confirmed she is confident that
she wants to go through with these changes (they're also in the [demo
designs](https://www.figma.com/design/ELhRhttKES1VaLkLNRLE5d/Demo-Mode?node-id=2047-17273&t=wCKXvY2LAPM0Q8Au-0))
- rationale being taking away some color will bring more emphasis to the
network icons.

### Test plan

![Simulator Screenshot - iPhone 15 Pro - 2025-01-24 at 16 54
16](https://github.com/user-attachments/assets/2ef69518-4a49-41c1-ba5b-4f35569ef715)
![Simulator Screenshot - iPhone 15 Pro - 2025-01-24 at 16 52
59](https://github.com/user-attachments/assets/a02bd913-4d6e-4e98-87ca-9ba025173e11)


### Related issues

- Fixes RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)