fix: EnterAmount wrong formatting for "1.234,5678" formatted numbers (#6472)
### Description
This PR again tries to fix an ongoing issue with formatting number for
TokenEnterAmount component when number format in phone settings is set
to `1.234,5678`.
The issue was introduced by [previous
attempt](https://github.com/valora-inc/wallet/pull/6428) to fix wrong
formatting for Swap flow specifically.

Previously, the issue was due to unformatting function only expecting
numbers that are formatted according to the phone's settings. This
caused an issue when on the Swap flow we set the value of "To" textfield
with `replaceAmount` function using the value from the quote API
response which was a JS-compatible number. This format differed for the
users that had format set in settings to `1.234,5678` and caused the
actual issue. But this introduced another issue with the wrong
formatting for when the numbers that exceed `1000`.

In order to fix this, the condition for whether there's a need to skip
the formatting was moved to actual function that is used when replacing
the value with the quote API response. I've also added some explanatory
comments to some functions for better clarity on what's each function
intention is.

### Test plan
Edited e2e tests for Send to test proper formatting of big amounts while
using `typeText` function which simulates actual typing instead of
`replaceText` (which is more of a paste).

This change doesn't revert the previous fix:


https://github.com/user-attachments/assets/a5b913eb-f47c-4a5d-9fcf-1c704376c177

### Related issues

- Relates to [this PR](https://github.com/valora-inc/wallet/pull/6428)
- Relates to the bug mentioned
[here](https://valora-app.slack.com/archives/C04B61SJ6DS/p1738341507462059)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)