feat: prevent actions that require private key in demo mode (#6469)
### Description

Unfortunately not all flows that access the private key go through the
same functions, so this PR adds a new blocking bottom sheet that is a
screen and navigates to it from:
1. `ensurePincode` - this is used from the flows that launch CAB / reset
pin / view recovery phrase
2. `sendPreparedTransaction` - this is used from all the transaction
flows, except for walletconnect
3. walletconnect saga, where we display the action request.

There is often UI that awaits the outcome of the transaction send flows
(spinners, navigation etc.) so I am returning the same result as if a
user dismissed the pincode screen during the flow.

### Test plan

Tested all flows to ensure the bottom sheet is correctly triggered, and
no extra errors are displayed.


https://github.com/user-attachments/assets/f376d1a7-a62c-43fa-b8a3-c598f81711d8



https://github.com/user-attachments/assets/eb817506-1cc1-4a8f-91c0-f9f3560e1455



https://github.com/user-attachments/assets/7bcdf0ee-00d9-465c-89c7-41f613669440



https://github.com/user-attachments/assets/5fbedfcd-7e2d-4b00-b61f-cec38788eb75



https://github.com/user-attachments/assets/55d26718-fa74-4406-a02f-257fdbc293c9


### Related issues

- Fixes RET-1304

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)