chore(deps): `react-native-keychain` upgrade + migrate to `@interaxyz` fork (#6470)
### Description

Migrate `react-native-keychain` to the `@interaxyz` fork in order to get
rid of the patch.

The patch changes are implemented in the fork:

https://github.com/oblador/react-native-keychain/compare/master...mobilestack-xyz:react-native-keychain:intera-fork

Since the current version (8.2.0) is outdated, it was also upgraded to
the latest one (9.2.2), which required some refactoring.

### Test plan

* CI
* Manually tested on iOS device 
* Manually tested on an Android emulator

Manual test plan:
* Start existing wallet after upgrade
* Display recovery phrase
* Sign transaction
* Restore wallet using a recovery phrase
* Display recovery phrase

The library migrates the secrets storage in Android from
`SharedPreferences` to `DataStore` in [version
9.2.0](https://github.com/oblador/react-native-keychain/releases/tag/v9.2.0).
While the migration is expected to be seamless, perhaps we should do as
much manual testing on devices as we can to ensure there is no major
issue.
### Related issues

- Related to RET-1308

### Backwards compatibility

Y

### Network scalability

NA