chore: change apps/example/index.ts to index.tsx for JSX capabilities (#54)
### Description
As the title says. In order to be able to define custom screens with
`<Navigator.Screen ... />` syntax we need index file to be in `.tsx`
format.

### Test plan
N/A

### Related issues

- Relates to
[RET-1311](https://linear.app/valora/issue/RET-1311/allow-defining-custom-screens)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)