chore(deps): migrate `react-native-fs` to `@divvi` fork (#6479)
### Description

Migrate `react-native-fs` to the `@divvi` fork in order to get rid of
the patch.

The patch changes are implemented in the fork:

https://github.com/itinance/react-native-fs/compare/master...divvixyz:react-native-fs:master

### Additional notes

This library is not maintained anymore, so we have to migrate to another
one.

Possible alternatives (to consider in the future, as they aren't a
drop-in replacement):

*
[expo-file-system](https://docs.expo.dev/versions/latest/sdk/filesystem/)
(requires adding expo, which looks [cumbersome for a bare RN
project](https://docs.expo.dev/bare/installing-expo-modules/))
*
[@dr.pogodin/react-native-fs](https://dr.pogodin.studio/docs/react-native-file-system)(fork
of the original library, the docs are saying it is a drop-in
replacement, but it doesn't work that way)

### Test plan

* CI
* Next release App Store submission (patch is adding the
`PrivacyInfo.xcprivacy` file)

### Related issues

- Related to RET-1308

### Backwards compatibility

Y

### Network scalability

NA