feat: add an ability to add provide custom screens (#58)
### Description
As per the title.

I've tried multiple approaches to get full type-safety with generics
with no success. It seems like to be fully type-safe – the whole flow
have to be consistent with the usage and inheritance of
`getAppConfig()`. This will probably be easier to implement once there
is a stable type-safe solution for [this navigation
task](https://linear.app/valora/issue/RET-1310/allow-navigating-to-specific-known-screens).

I've also added an example component utilizing `useWallet` hook for
better clarity on how to add custom screens and how to use the new
exposed data hooks.

### Test plan
Example of custom screen using data from the `useWallet` hook:

<img
src="https://github.com/user-attachments/assets/24a9fdf5-484b-4856-945e-d0b1ff3c95bb"
width="400" />




### Related issues

- Relates to
[RET-1311](https://linear.app/valora/issue/RET-1311/allow-defining-custom-screens)

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)