fix(format): Prevent prettier from formatting RootStateSchema (#61)
### Description

When running `yarn format` from within the `packages/@divvi/mobile/`
directory, prettier does not see/use the `.prettierignore` in the repo
top-level directory. Adding this in prevents `yarn format` from mangling
`RootStateSchema.json` in particular.

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)