refactor: remove multichain feature dynamic config (#6487)
### Description

This PR removes the MULTICHAIN_FEATURES Statsig dynamic config variable.
It seems reasonable to assume that builders would not need the ability
to _dynamically_ change the network their app supports. It also seems
reasonable to assume that builders would want to support a whole network
fully, and thus the ability to control which networks have access to tx
feed / cico / approval transactions etc. seems unnecessary. With these
assumptions, the supported networks seem like they don't need to go
through Statsig at all.

This PR:
1. move the supported networks to config.ts but eventually this will be
app config that is passed in via the `createApp` config. access the
network ids via a central function (where we can read from config and
default to all networks if not provided)
2. remove MULTICHAIN_FEATURES dynamic config and util functions that
gate features by network (replacing all usages with a single
`getSupportedNetworkIds`)
3. update tests

The MULTICHAIN_FEATURES config was used when adding new networks,
however I assume that a different approach will be used if / when adding
new networks going forward since we would do this on the framework layer
(only allowing builders to access the new network once all
functionalities are supported)

### Test plan

n/a

### Related issues

- RET-1319

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)