chore: add redux caching for divvi registration status (#6490)
### Description

This PR adds the redux variable / actions to track the divvi
registration status. It will be used in the next PRs. It is not strictly
necessary but does reduce the need to fetch the status on-chain every
transaction.

See #6485 for more context of how it will be used.

### Test plan

No impact on existing functionality.

### Related issues

- Related to RET-1318

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)