refactor: ensure all transaction flows can handle registration transactions (#6491)
### Description

We plan to add registration transactions to `prepareTransactions`, but
some transaction flows have logic that assumes all prepared transactions
are related to the flows. This PR modifies key transaction sending flows
to handle prepared transactions that aren't directly related to their
flows by filtering out the registration transactions before proceeding
with their logic.

It's not particularly nice for every flow to know about registrations,
but i think the crux of the issue is that the flows think that they know
what is inside the prepared transactions :/

There are some flows that require tweaks to support registrations (for
various reasons, one being that some flows only pass along the first
prepared transaction rather than all the prepared transactions). The
send / walletconnect / fiatconnect / live links reclaim / cash in / cash
out flows are not yet supported.

### Test plan

This should have no user facing changes.

### Related issues

- Related to RET-1318

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)