fix: TS typecheck not working in the example app (#79)
### Description

TS typecheck was working but actually failing because it only included
the `src` folder which is not present in the example.

I banged my head a little trying to make it all work (things work
slightly differently when typechecking in the context of the library vs
the example app).

The main problem is we don't currently ship the types with the mobile
library, requiring some specific setup in `tsconfig.json` of the final
app.

I tried emitting the declarations but it gave too many errors (1400+),
mostly similar to:

```
Exported variable 'defaultCountryCodeSelector' has or is using name 'State' from external module "/Users/jean/src/github.com/divvixyz/divvi-mobile/packages/@divvi/mobile/src/account/reducer" but cannot be named.
```

I think we'll be able to solve this once the wallet is consuming the
library and we can make bigger changes in this repo (relative imports,
require cycles, etc).

In the meantime, the current changes are much more limited in scope and
allow TS checks to pass properly.
I've also added a CI check so we don't break them.