chore: handle referrer not registered in divvi integration (#6496)
### Description

This PR ensures that the registration tx only contains protocolId's that
the referrer is registered for. This prevents the prepared transaction
from containing a protocolId that would cause the tx to revert (thereby
blocking the rest of the transaction flow). A builder could be in this
situation if they are developing before they are added to the
registration contract.

### Test plan

Setting the referrer id to a random string or adding a random protocol
id to the app config does not affect the transaction flow.

### Related issues

- Fixes RET-1318

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Jean Regisser <jean.regisser@gmail.com>