chore(ui): refactor `SkeletonPlaceholder` usage (#6483)
### Description

We want to get rid of `react-native-skeleton-placeholder` patch which
whole purpose is to allow `testID` prop on the skeleton component. Since
it is not clear when the PR adding this patch to the original library
will be released, this PR aims to provide similar functionality by
wrapping the original skeleton component into a `View`.

### Additional refactoring

* Standardized imports: all Skeleton use cases import it from a custom
component, not directly from the library
* Standardized defaults: `borderRadius`, `backgroundColor`,
`highlightColor` have reasonable default values (can be overridden if
needed)

### Test plan

* CI
* Tested manually on Celo News, Claim Earn, TokenEnterAmount, NFT
components

### Related issues

- Fixes RET-1308

### Backwards compatibility

Y

### Network scalability

NA