chore: prepare to divvi migration (#6529)
### Description

Delete redundant artifacts from the repo to facilitate migration to the
[divvi mobile
framework](https://github.com/divvi-xyz/divvi-mobile/tree/main).

This PR contains only entire file deletions, no other changes, to make
it simpler to review.

This is the first part of the migration. On top of this cleanup, we will
submit the actual app configuration in the immediate subsequent PR:
https://github.com/valora-inc/wallet/pull/6530

### Deletion reasons

We expect we will not need most of the files being deleted in this repo
anymore.

| Deleted entity | Reason for deleting |
|--------|--------|
| `locales` directory, Crowdin config, and Crowdin-related Github
actions | Translations are updated within the framework now. We don't
yet have any established way to update custom translations in the
consumer app automatically |
| Release-related Github actions and `fastlane` directory | We will
migrate to EAS for releases similar to the [Beefy
app](https://github.com/valora-inc/beefy/pull/10). Note the
`release-nightly.yml` isn't deleted, since we are planning to reuse it
with EAS |
| Update licenses Github action | Licenses are updated within the
framework. We don't have yet any established way to update licenses in
the consumer app, if any |
| `__mocks__`, `e2e`, and `test` directories | The divvi framework is
covered with the same tests as the wallet. We don't have yet any unit
tests or e2e tests ready for the consumer app. We plan to establish a
way to create unit tests in the consumer apps shortly. See RET-1326 |
| `android`, `ios` directories, and `Gemfile` | It will be
code-generated by Expo, and not required to be checked in the repo
anymore |
| `patches` directory | Currently applied patches become redundant after
migrating from the bare React Native to Expo |
| `src` directory | All of it is the divvi framework now. The current
approach to structuring consumer apps doesn't require an `src` directory
|
| `react-native.config.js` file | Assets are managed differently in Expo
|
| `app.json` file | We will use `app.config.js` instead to have more
flexibility for different app flavors (mainnet, nightly, alfajores) |
| `index.js` file | We will have a different entry point with the
framework |

### Test plan

This PR is planned to be merged without CI, nothing to test here

### Related issues

- Related to RET-1321