chore: update renovate config (#6555)
### Description

Update renovate config:
- automatically update `@divvi/mobile` (following the alpha tag)
whenever it updates
- don't update native dependencies, as they need to be compatible with
the version used by the framework

This is similar to https://github.com/divvi-xyz/divvi-app-starter/pull/2

The existing config (before this PR) could potentially update to a major
version that wouldn't actually work with the app at runtime.
Since the e2e test here just compiles the app and doesn't yet run it.

We'll update native dependencies in the framework and we'll provide a
solution for consuming apps to follow them without manual work.

### Test plan

N/A

### Related issues

N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)