fix: always use 'Valora' for registryName and User-Agent (#6592)
### Description

One of the loose ends of the config, the name used for the `User-Agent`
(which defaults to the expo app name) should be the same as the
`registryName`.
But these configs are set in different places (config plugin vs
createApp).
So here we pass it to both places, so they are consistent.

Basically this fixes potential issues with dev / nightly builds not
behaving like the prod build with the `User-Agent`