chore: fix release notes script (#6624)
### Description

This fixes `yarn generate-release-notes` failing.
Also added a new CI check to ensure it keeps working in the future.

See Slack
[thread](https://valora-app.slack.com/archives/C025X4WJT09/p1743784034647079?thread_ts=1743710719.473429&cid=C025X4WJT09).

### Test plan

- CI passes
- Manually confirmed `yarn generate-release-notes` now works

### Related issues

N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)